chore: Update .eslintignore, .gitignore, and package.json
This commit updates the .eslintignore file to ignore the "build" and "node_modules" directories. It also updates the .gitignore file to include the "database/*.db" files, ".env", and ".env.test" files. Additionally, it adds a new "build" script to the package.json file to compile the TypeScript code into the "build" directory. These changes improve the project's development and build process by properly ignoring unnecessary files and directories and adding a build script for easier compilation.